backend/requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-anthropic>=1.0.0
langchain-openai>=0.2.0
pandas>=2.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
firebase-admin>=6.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
pydantic[email]>=2.0.0
cryptography>=42.0.0
markdown>=3.5.0
weasyprint>=62.0